refectoring dec, dmm, dms
